refactor: unify test helpers
