+++ docs/developer_guide.md
@@ -185,7 +185,7 @@ Sample archives used by the tests are versioned in `tests/test_archives`.  If
 you add new archives or change them, regenerate the files with:
 
 ```bash
-uv run --extra optional python -m tests.create_archives [pattern]
+uv run --extra optional python -m tests.archivey.create_archives [pattern]
 ```
 
 Omit the optional pattern to rebuild all archives.  RAR tests require the
